+++ python/legends_firecrawl/__init__.py
@@ -2,7 +2,7 @@
 
 from .client import ApiError, CredentialError, FirecrawlClient, SafetyError
 
-__version__ = "0.2.1"
+__version__ = "0.1.0"
 
 __all__ = [
     "ApiError",